security

+++ RAMS/Web/RAMMS.Domain/Models/RmB11MaterialCost.cs
@@ -11,7 +11,7 @@ namespace RAMMS.Domain.Models
         public string B11mcMaterialId { get; set; }
         public string B11mcMaterialName { get; set; }
         public decimal? B11mcMaterialPerUnitPrice { get; set; }
-        public int? B11mcMaterialNoOfUnits { get; set; }
+        public decimal? B11mcMaterialNoOfUnits { get; set; }
         public decimal? B11mcMaterialTotalPrice { get; set; }
         public int? B11mcMaterialOrderId { get; set; }
